+++ composeApp/src/commonMain/kotlin/com/mnessim/rsstracker/Constants.kt
@@ -1,11 +1,7 @@
 package com.mnessim.rsstracker
 
 object Constants {
-    const val APP_VERSION: String = "1.0.11"
-
-    //    const val RUST_BACKEND_URL: String = "http://localhost:3000"
+    const val APP_VERSION: String = "1.0.13"
     const val RUST_BACKEND_URL: String = "https://rust.mnessim.com"
-
-    //    const val DEV_ENV: Boolean = true
     const val APP_TITLE: String = "RssTracker"
 }